+++ src/main/java/command/ExitCommand.java
@@ -1,0 +1,15 @@
+package command;
+
+import storage.Storage;
+
+import task.TaskList;
+
+import ui.Ui;
+
+public class ExitCommand extends Command {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showLine();
+        System.out.println("Goodbye!");
+    }
+    public boolean isExit() { return true; }
+}
