+++ src/main/java/dew/DewException.java
@@ -1,3 +1,5 @@
+package dew;
+
 public class DewException extends Exception {
     public DewException(String message) {
         super("____________________________________________________________\n"
