+++ src/main/java/task/Deadline.java
@@ -1,5 +1,6 @@
+package task;
+
 import java.time.LocalDate;
-import java.util.Locale;
 
 public class Deadline extends Task {
