+++ src/main/java/task/Event.java
@@ -1,4 +1,6 @@
-public class Event extends  Task{
+package task;
+
+public class Event extends Task {
 
     protected String timeStart;
     protected String timeEnd;
