+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ Hello from
 |____/ \___|  \__/   \__/
 
 ____________________________________________________________
- Hello! I'm Dew
+ Hello! I'm dew.Dew
  What can I do for you?
 ____________________________________________________________
